renames gcp to bigquery client
