+++ dlt/pipeline/typing.py
@@ -6,7 +6,7 @@ from dlt.common import json
 from dlt.common.typing import StrAny
 from dlt.common.configuration.utils import TSecretValue
 
-TLoaderType = Literal["gcp", "redshift"]
+TLoaderType = Literal["bigquery", "redshift"]
 TPipelineStage = Literal["extract", "normalize", "load"]
 
 # extractor generator yields functions that returns list of items of the type (table) when called
@@ -49,7 +49,7 @@ class GCPPipelineCredentials(PipelineCredentials):
     def from_services_dict(cls, services: StrAny, dataset_prefix: str) -> "GCPPipelineCredentials":
         assert dataset_prefix is not None
 
-        return cls("gcp", services["project_id"], dataset_prefix, services["client_email"], services["private_key"])
+        return cls("bigquery", services["project_id"], dataset_prefix, services["client_email"], services["private_key"])
 
     @classmethod
     def from_services_file(cls, services_path: str, dataset_prefix: str) -> "GCPPipelineCredentials":
