+++ tests/loaders/bigquery/test_bigquery_client.py
@@ -10,7 +10,7 @@ from dlt.common.utils import uniq_id
 from dlt.load.exceptions import LoadJobNotExistsException, LoadJobServerTerminalException
 
 from dlt.load import Load
-from dlt.load.gcp.client import BigQueryClient
+from dlt.load.bigquery.client import BigQueryClient
 
 from tests.utils import TEST_STORAGE, delete_storage
 from tests.loaders.utils import expect_load_file, prepare_event_user_table, yield_client_with_storage
@@ -18,7 +18,7 @@ from tests.loaders.utils import expect_load_file, prepare_event_user_table, yiel
 
 @pytest.fixture(scope="module")
 def client() -> Iterator[BigQueryClient]:
-    yield from yield_client_with_storage("gcp")
+    yield from yield_client_with_storage("bigquery")
 
 
 @pytest.fixture
@@ -34,7 +34,7 @@ def auto_delete_storage() -> None:
 def test_empty_schema_name_init_storage(client: BigQueryClient) -> None:
     e_client: BigQueryClient = None
     # will reuse same configuration
-    with Load.import_client_cls("gcp", initial_values={"DEFAULT_DATASET": client.CONFIG.DEFAULT_DATASET})(Schema("")) as e_client:
+    with Load.import_client_cls("bigquery", initial_values={"DEFAULT_DATASET": client.CONFIG.DEFAULT_DATASET})(Schema("")) as e_client:
         e_client.initialize_storage()
         try:
             # schema was created with the name of just schema prefix
